Clear all sections of color when updating the map
When we update available ticket groups, we're only looping through the new
set of available ticket groups when updating the section colors on the map.
We aren't taking into consideration sections which were previously available
but no longer available. Those sections need to be marked unavailable before
we loop through the new collection to highlight.